TG-52: create team-members-table shared component
