+++ src/Components/Share/SelectForm/selectForm.module.css
@@ -15,7 +15,7 @@
 .flexLabel,
 .flexImput,
 .flexSelect {
-  width: 85%;
+  width: 100%;
   color: var(--color-blue-primary);
   cursor: pointer;
 }
